requirements.txt
streamlit
google-generativeai
python-dotenv
google-ai-generativelanguage
grpcio